+++ src/main/java/gg/dsmedia/megafloaters/MegaFloatersMod.java
@@ -4,6 +4,7 @@ import com.mojang.logging.LogUtils;
 import gg.dsmedia.megafloaters.command.MegaFloatersCommand;
 import gg.dsmedia.megafloaters.event.DiscoveryTracker;
 import gg.dsmedia.megafloaters.event.ObservationHooks;
+import gg.dsmedia.megafloaters.event.ReloadListeners;
 import gg.dsmedia.megafloaters.event.SpawnSuppression;
 import gg.dsmedia.megafloaters.integration.AeronauticsCompat;
 import gg.dsmedia.megafloaters.integration.BddCompat;
@@ -27,6 +28,7 @@ public class MegaFloatersMod {
         NeoForge.EVENT_BUS.addListener(SpawnSuppression::onFinalizeSpawn);
         NeoForge.EVENT_BUS.addListener(DiscoveryTracker::onPlayerTick);
         NeoForge.EVENT_BUS.addListener(ObservationHooks::onRightClickBlock);
+        NeoForge.EVENT_BUS.addListener(ReloadListeners::onAddReloadListeners);
         LOGGER.info("Mega Floaters initialising.");
     }